moved time_now to client, renamed time_now_128
